I have done migration of db to new data type of amount in each table

+++ backend/Zlotowka/Zlotowka.Server/Models/SpendRequest.cs
@@ -3,6 +3,6 @@
 public class SpendRequest
 {
     public string SpendTitle { get; set; }
-    public int SpendAmount { get; set; }
+    public decimal SpendAmount { get; set; }
     public DateTime SpendDate { get; set; }
 }
